Added support for position as string
It seems that in some cases the getPositionAccessor function may return a string instead of an object. From the `movement_pittsburgh` example it seems that it happens with CSV files, but I'm not 100% sure if that's always the case.

In any case, it shouldn't hurt to try and parse the string value, and it works just fine in the scenario described above.